+++ tiger-rbel/src/main/java/de/gematik/rbellogger/util/RbelContent.java
@@ -485,4 +485,8 @@ public class RbelContent {
     }
     return foundCount; // fewer than max allowed found
   }
+
+  public boolean contains(byte[] searchContent) {
+    return indexOf(searchContent) >= 0;
+  }
 }